backend/requirements.txt
#
# AiMediaOps 依赖清单
# 说明：
#   1. 仅包含项目实际使用的高层依赖，不包含 venv 中的所有包
#   2. 请使用 `python -m venv venv` 创建虚拟环境，并在其中安装本文件列出的依赖
#   3. venv/.venv 已在 .gitignore 中忽略，请不要提交到 Git
#

mcp

# ==============================
# 核心依赖 - Core Dependencies
# ==============================
pydantic>=2.0.0
jinja2>=3.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
tenacity>=8.4.0

# ==============================
# 异步支持 - Async Support
# ==============================
# asyncio-py>=3.11.0  # asyncio 是 Python 3.7+ 的标准库，不需要安装
aiofiles>=23.2.0
aiohttp>=3.9.0

# ==============================
# Web API 服务 - FastAPI
# ==============================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# ==============================
# 注册码与加密 - License & Security
# ==============================
cryptography>=43.0.0
httpx>=0.26.0

# ==============================
# 开发工具 - Development Tools
# ==============================
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# ==============================
# 性能监控 - Performance Monitoring
# ==============================
psutil>=5.9.0
memory-profiler>=0.61.0

# ==============================
# 类型检查 - Type Checking
# ==============================
types-requests>=2.31.0

# ==============================
# 日志和配置 - Logging and Configuration
# ==============================
pyyaml>=6.0
python-box>=7.0.0
click>=8.0.0
rich>=13.0.0
loguru>=0.7.0
diskcache>=5.6.0

# ==============================
# 浏览器自动化 - Browser Automation
# ==============================
playwright>=1.40.0
# browser-use>=0.2.0  # 可选，如果使用 browser-use

# ==============================
# 其他说明
# ==============================
# 1. 不要把 venv/.venv 提交到 Git，已经在 .gitignore 中排除
# 2. 安装 playwright 后需要执行: playwright install chromium

